#8200 - move Issue8200Test to ClassTemplateTest
